Major updates to the race datastore.
Adds a few races and a few columns. This change transitions to removing
any racial abilities that are not conditional or active. Abilities like
keen-senses are just a +2 racial bonus to perception, so we don't want
to print those on the abilities section of the character sheet, and
those are instead captured in the skills column.